+++ src/JsonApi/Response/AbstractResponder.php
@@ -138,8 +138,8 @@ abstract class AbstractResponder
         }
 
         $content = $document->getContent($additionalMeta);
-        $responseCode = $document->getResponseCode($statusCode);
+        $statusCode = $document->getStatusCode($statusCode);
 
-        return $this->serializer->serialize($this->response, $responseCode, $content);
+        return $this->serializer->serialize($this->response, $statusCode, $content);
     }
 }