+++ CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGED:
 - `AbstractResource::getLinks()` returns `?ResourceLinks` instead of `Links` (__BREAKING__)
 - `AbstractRelationship::getLinks()` returns `?RelationshipLinks` instead of `Links` (__BREAKING__)
 - `AbstractRelationship::setLinks()` expects a parameter of `?RelationshipLinks` type instead of `Links` (__BREAKING__)
+- Renamed `AbstractErrorDocument::getResponseCode()` to `AbstractErrorDocument::getStatusCode()` (__BREAKING__)
 
 REMOVED: